system/backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic-settings
python-multipart
passlib[bcrypt]
bcrypt==4.0.1
email-validator
pandas
requests
